texture/geo2rgb/test_controlnet_6views/requirements1.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
carvekit==4.1.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.108.0
filelock==3.15.1
frozenlist==1.4.1
fsspec==2024.6.0
fvcore==0.1.5.post20221221
h11==0.14.0
idna==3.4
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
loguru==0.7.2
MarkupSafe==2.1.5
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
Pillow==10.1.0
portalocker==2.8.2
psutil==5.9.8
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.5.0
sniffio==1.3.1
starlette==0.32.0.post1
sympy==1.12.1
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.1.2+cu118
torch_cluster==1.6.3
torch_geometric==2.5.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.4
triton==2.1.0
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==1.26.13
uvicorn==0.25.0
yacs==0.1.8
yarl==1.9.4
absl-py
accelerate
addict
aiobotocore
aioitertools
albumentations
alembic
altair
aniso8601
antlr4-python3-runtime
appdirs
apptools
arrow
asgiref
asttokens
backcall
backoff
beartype
beautifulsoup4
blessed
blinker
boto3
botocore
brotlipy
build
cachetools
cffi
cholespy
chumpy
clip
cloudpickle
cmake
CoCa-pytorch
coloredlogs
comm
ConfigArgParse
configobj
contourpy
cos-python-sdk-v5
crcmod
croniter
cycler
Cython
dash
dash-core-components
dash-html-components
dash-table
datasets
dateutils
debugpy
decorator
deepdiff
diffusers
dill
dlib @ file:///aigc_cfs_2/neoshang/code/dlib/dist/dlib-19.24.99-cp39-cp39-linux_x86_64.whl#sha256=ee6cd29ef4d7d0b66163064068bc95b24ac46c6469feda5e0b8f000086f8ceec
dnspython
docker-pycreds
dpcpp-cpp-rt
easydict
einops
einops-exts
ema-pytorch
email_validator
envisage
executing
face-alignment
fastapi-cli
fastjsonschema
ffmpy
fire
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
flask-swagger-ui
flatbuffers
fonttools
ftfy
future
gdown
gevent
gitdb
GitPython
glfw
google-auth
google-auth-oauthlib
greenlet
grpcio
grpcio-tools
h5py
httpcore
httptools
httpx
huggingface-hub
humanfriendly
imageio
imageio-ffmpeg
importlib-metadata
importlib-resources
iniconfig
inquirer
intel-cmplr-lib-rt
intel-cmplr-lic-rt
intel-opencl-rt
intel-openmp
invisible-watermark
ipdb
ipykernel
ipython
ipywidgets
itsdangerous
jaxtyping
jedi
jmespath
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab-widgets
kaleido
kiui
kiwisolver
kornia
lazy_loader
libigl
lightning-api-access
lightning-cloud
lightning-utilities
linkify-it-py
lit
llvmlite
Mako
Markdown
markdown-it-py
matplotlib
matplotlib-inline
mayavi
mdit-py-plugins
mdurl
mkl
multiprocess
mypy-extensions
natsort
nbformat
nest-asyncio
ninja
numba
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast/@335cfa6b33d785730a04283994214bed57884e87
oauthlib
objprint
omegaconf
onnxruntime
open-clip-torch
open3d
opencv-python-headless
ordered-set
orjson
packaging
pandas
parso
pathtools
pexpect
pickleshare
platformdirs
plotly
pluggy
plumbum
plyfile
pooch
prompt-toolkit
protobuf
ptyprocess
pudb
pulsar-client
pure-eval
pyarrow
pyarrow-hotfix
pyasn1
pyasn1-modules
pycparser
pycryptodome
pydeck
pyDeprecate
pydub
pyface
PyGLM
Pygments
PyJWT
PyMatting
PyMCubes
pymeshlab
Pympler
PyMySQL
PyOpenGL
pyproject_hooks
PyQt5
PyQt5-Qt5
PyQt5-sip
pyquaternion
pyre-extensions
pyrsistent
PySocks
pytest
python-dateutil
python-dotenv
python-editor
pytorch-lightning
pytorch-warmup
pytz
pytz-deprecation-shim
PyWavelets
pyzmq
qudida
readchar
redis
referencing
regex
rembg
requests-oauthlib
resize-right
responses
rich
rotary-embedding-torch
rpds-py
rpyc
rsa
s3fs
s3transfer
safetensors
scikit-image
segment-anything
semantic-version
sentencepiece
sentry-sdk
setproctitle
shellingham
six
smmap
smplx
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
starsessions
submitit
sympy
tabulate
taming-transformers @ git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318
tbb
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
termcolor
test_tube
threadpoolctl
tifffile
timm
tokenizers
toml
tomli
toolz

torch-fidelity
torch-sparse
torch-tb-profiler
torchmetrics==0.10.3
tornado
tqdm
traitlets
traits
traitsui
transformers
trimesh
triton
typeguard
typer
typing
typing-inspect
typing_extensions
tzdata
tzlocal
uc-micro-py
ujson
urllib3
urwid
urwid_readline
uvicorn
uvloop
validators
varname
vector-quantize-pytorch
vtk
wandb
watchdog
watchfiles
wcwidth
websocket-client
websockets
Werkzeug
widgetsnbextension
wrapt
xatlas
xmltodict
xxhash
yacs
yarl
zipp
zope.event
zope.interface